requirements.txt
aiohttp==3.5.4
aioredis==1.2.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
chardet==3.0.4
hiredis==1.0.0
idna==2.8
more-itertools==6.0.0
multidict==4.5.2
nltk==3.4
pluggy==0.9.0
py==1.8.0
pytest==4.3.0
pytest-aiohttp==0.3.0
PyYAML==3.13
singledispatch==3.4.0.3
six==1.12.0
trafaret==1.2.0
yarl==1.3.0
